+++ source/CHANGELOG.md
@@ -1,7 +1,19 @@
-## 0.0.43
+## 0.0.44
 added methods:
 - `onFirstMonthWeekDay` (shortcut for `dom:1W`)
 - describe works (only english locale shipped)
+
+
+## 0.0.43  
+cadences optional added in:  
+- `atSecond`, `atSecondAdd`
+- `atMinute`, `atMinuteAdd`
+- `atHour`, `atHourAdd`
+- `atWeekDay`, `atWeekDayAdd`
+- `atMonthDay`, `atMonthDayAdd`
+- `atMonth`, `atMonthAdd`
+- `atYear`, `atYearAdd`
+
 
 ## 0.0.42  
 - fixed wrong behaviour of `betweenWeekDays`